+++ custom-addons/esb/account/models/__init__.py
@@ -1,3 +1,4 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html)
 
 from . import account_move
+from . import withholding_tax_cert
