+++ custom-addons/purchase_sale_inter_company/__manifest__.py
@@ -6,7 +6,7 @@
 {
     "name": "Inter Company Module for Purchase to Sale Order",
     "summary": "Intercompany PO/SO rules",
-    "version": "13.0.1.0.5",
+    "version": "13.0.1.0.0",
     "category": "Purchase Management",
     "website": "http://www.github.com/OCA/multi-company",
     "author": "Odoo SA, Akretion, Tecnativa, Odoo Community Association (OCA)",
