+++ custom-addons/purchase_sale_inter_company/i18n/fr_FR.po
@@ -3,26 +3,26 @@
 # * purchase_sale_inter_company
 #
 # Translators:
-# Matjaž Mozetič <m.mozetic@matmoz.si>, 2017
+# OCA Transbot <transbot@odoo-community.org>, 2017
 msgid ""
 msgstr ""
 "Project-Id-Version: Odoo Server 8.0\n"
 "Report-Msgid-Bugs-To: \n"
 "POT-Creation-Date: 2017-10-19 00:51+0000\n"
 "PO-Revision-Date: 2017-10-19 00:51+0000\n"
-"Last-Translator: Matjaž Mozetič <m.mozetic@matmoz.si>, 2017\n"
-"Language-Team: Slovenian (https://www.transifex.com/oca/teams/23907/sl/)\n"
-"Language: sl\n"
+"Last-Translator: OCA Transbot <transbot@odoo-community.org>, 2017\n"
+"Language-Team: French (France) (https://www.transifex.com/oca/teams/23907/"
+"fr_FR/)\n"
+"Language: fr_FR\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
-"Plural-Forms: nplurals=4; plural=(n%100==1 ? 0 : n%100==2 ? 1 : n%100==3 || n"
-"%100==4 ? 2 : 3);\n"
+"Plural-Forms: nplurals=2; plural=(n > 1);\n"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model,name:purchase_sale_inter_company.model_res_company
 msgid "Companies"
-msgstr "Družbe"
+msgstr "Sociétés"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model,name:purchase_sale_inter_company.model_res_config_settings
@@ -41,8 +41,8 @@ msgid ""
 "Default value to set on Sale Orders that will be created based on Purchase "
 "Orders made to this company"
 msgstr ""
-"Privzeta vrednost nastavljena na prodajnih nalogih, ki bo ustvarjena na "
-"podlagi nabavnih nalogov za to družbo."
+"Valeur par défaut mise sur les commandes de vente créés à partir des "
+"commandes d'achat faite à cette société"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,help:purchase_sale_inter_company.field_res_config_settings__warehouse_id
@@ -50,8 +50,8 @@ msgid ""
 "Default value to set on Sale Orders that will be created based on Purchase "
 "Orders made to this company."
 msgstr ""
-"Privzeta vrednost nastavljena na prodajnih nalogih, ki bo ustvarjena na "
-"podlagi nabavnih nalogov za to družbo."
+"Valeur par défaut mise sur les commandes de vente créés à partir des "
+"commandes d'achat faite à cette société."
 
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,help:purchase_sale_inter_company.field_res_company__so_from_po
@@ -81,19 +81,20 @@ msgstr ""
 #. module: purchase_sale_inter_company
 #: model:ir.model,name:purchase_sale_inter_company.model_purchase_order
 msgid "Purchase Order"
-msgstr "Nabavni nalog"
+msgstr "Commande d'achat"
 
 #. module: purchase_sale_inter_company
 #: model_terms:ir.ui.view,arch_db:purchase_sale_inter_company.res_config_settings_view_form
 #, fuzzy
+#| msgid "Purchase to Sale"
 msgid "Purchase/Sale"
-msgstr "Nabava v prodajo"
+msgstr "Achat vers Vente"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,field_description:purchase_sale_inter_company.field_res_company__sale_auto_validation
 #: model:ir.model.fields,field_description:purchase_sale_inter_company.field_res_config_settings__sale_auto_validation
 msgid "Sale Orders Auto Validation"
-msgstr "Samodejno overjanje prodajnih nalogov"
+msgstr "Auto Validation des commandes de vente"
 
 #. module: purchase_sale_inter_company
 #: model_terms:ir.ui.view,arch_db:purchase_sale_inter_company.res_config_settings_view_form
@@ -103,23 +104,24 @@ msgstr ""
 #. module: purchase_sale_inter_company
 #: model:ir.model,name:purchase_sale_inter_company.model_sale_order
 #, fuzzy
+#| msgid "Sales Order"
 msgid "Sales Order"
-msgstr "Prodajni nalog"
+msgstr "Commandes de vente"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model,name:purchase_sale_inter_company.model_sale_order_line
 msgid "Sales Order Line"
-msgstr "Postavka prodajnega naloga"
+msgstr "Ligne de commandes de vente"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,field_description:purchase_sale_inter_company.field_sale_order__auto_purchase_order_id
 msgid "Source Purchase Order"
-msgstr "Izvorni nabavni nalog"
+msgstr "Commande d'achat source"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,field_description:purchase_sale_inter_company.field_sale_order_line__auto_purchase_line_id
 msgid "Source Purchase Order Line"
-msgstr "Postavka izvornega prodajnega naloga"
+msgstr "Ligne de commande d'achat source"
 
 #. module: purchase_sale_inter_company
 #: code:addons/purchase_sale_inter_company/models/stock_picking.py:0
@@ -150,13 +152,14 @@ msgstr ""
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,field_description:purchase_sale_inter_company.field_res_company__warehouse_id
 msgid "Warehouse For Sale Orders"
-msgstr "Skladišče za prodajne naloge"
+msgstr "Entrepôt pour les commandes de vente"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,field_description:purchase_sale_inter_company.field_res_config_settings__warehouse_id
 #, fuzzy
+#| msgid "Warehouse For Sale Orders"
 msgid "Warehouse for Sale Orders"
-msgstr "Skladišče za prodajne naloge"
+msgstr "Entrepôt pour les commandes de vente"
 
 #. module: purchase_sale_inter_company
 #: model:ir.model.fields,help:purchase_sale_inter_company.field_res_company__sale_auto_validation
@@ -165,8 +168,8 @@ msgid ""
 "When a Sale Order is created by a multi company rule for this company, it "
 "will automatically validate it."
 msgstr ""
-"Ko se ustvari prodajni nalog iz pravila konsolidiranja za to družbo, se le-"
-"ta samodejno overi."
+"Quand une commande de vente est créé en mode inter-société pour cette "
+"société, elle sera automatiquement validé."
 
 #. module: purchase_sale_inter_company
 #: code:addons/purchase_sale_inter_company/models/purchase_order.py:0
@@ -179,50 +182,53 @@ msgstr ""
 #, python-format
 msgid "You cannot create SO from PO because product '%s' is not intercompany"
 msgstr ""
-"Prodajnega naloga ni mogoče ustvariti na osnovi nabavnega, ker proizvoda "
-"'%s' ni v sklopu strukturiranih družb."
+"Vous ne pouvez pas créer la SO à partir de la PO car le produit '%s' n'est "
+"pas multi-société"
 
 #. module: purchase_sale_inter_company
 #: code:addons/purchase_sale_inter_company/models/purchase_order.py:0
 #, fuzzy, python-format
+#| msgid ""
+#| "You cannot create SO from PO because sale price list currency is "
+#| "different from purchase price list currency."
 msgid ""
 "You cannot create SO from PO because sale price list currency is different "
 "than purchase price list currency."
 msgstr ""
-"Prodajnega naloga ni mogoče ustvariti na osnovi nabavnega, ker je valuta "
-"prodajnega cenika drugačna od valute nabavnega."
+"Vous ne pouvez pas créer la SO à partir de la PO car la devise sur la liste "
+"de prix de vente est différente de la devise sur la liste de prix d'achat."
 
 #~ msgid "Invoice"
-#~ msgstr "Račun"
+#~ msgstr "Facture"
 
 #~ msgid ""
 #~ "Configure correct warehouse for company (%s) in Menu: Settings/companies/"
 #~ "companies"
 #~ msgstr ""
-#~ "Nastavite pravilno skladišče za družbo (%s) v meniju: Nastavitve/Družbe/"
-#~ "Družbe"
+#~ "Configurez correctement l'entrepôt pour la société (%s) dans le Menu: "
+#~ "Configuration/sociétés/sociétés"
 
 #~ msgid "Output - A"
-#~ msgstr "Izhod - A"
+#~ msgstr "Output - A"
 
 #~ msgid "Output - B"
-#~ msgstr "Izhod - B"
+#~ msgstr "Output - B"
 
 #~ msgid "Sale Auto Validation"
-#~ msgstr "Samodejno overjanje prodaje"
+#~ msgstr "Auto Validation de la commande de vente"
 
 #~ msgid "Stock - A"
-#~ msgstr "Zaloga - A"
+#~ msgstr "Stock - A"
 
 #~ msgid "Stock - B"
-#~ msgstr "Zaloga - B"
+#~ msgstr "Stock - B"
 
 #~ msgid ""
 #~ "When a Sale Order is created by a multi company rule for this company, it "
 #~ "will automatically validate it"
 #~ msgstr ""
-#~ "Ko se ustvari prodajni nalog iz pravila konsolidiranja za to družbo, se "
-#~ "le-ta samodejno overi"
+#~ "Quand une commande de vente est créé en mode inter-société pour cette "
+#~ "société, elle sera automatiquement validé"
 
 #~ msgid ""
 #~ "{'invisible':['|', ('invoice_method','in', ['picking', 'manual', "
