+++ custom-addons/purchase_sale_inter_company/models/__init__.py
@@ -3,4 +3,3 @@ from . import purchase_order
 from . import res_company
 from . import res_config
 from . import sale_order
-from . import stock_picking
