+++ custom-addons/purchase_sale_inter_company/models/res_config.py
