+++ custom-addons/purchase_sale_inter_company/readme/CONFIGURE.rst
@@ -1,0 +1,7 @@
+To configure this module, you need to:
+
+#. go to the menu *Settings > Companies > Companies*.
+#. Select one of the companies.
+#. Go to the tab *Inter-Company* then the group *Purchase To Sale*.
+#. Select the *Warehouse For Sale Orders*, it is the warehouse that will be used to automatically generate the sale order in the other company.
+#. If you check the option *Sale Auto Validation* in the configuration of company B, then when you validate a *Purchase Order* in company A with company B as supplier, the *Sale Order* will be automatically validated in company B with company A as customer.
