+++ custom-addons/purchase_sale_inter_company/readme/CONTRIBUTORS.rst
@@ -1,0 +1,9 @@
+* Odoo S.A.
+* Chafique Delli <chafique.delli@akretion.com>
+* Alexis de Lattre <alexis.delattre@akretion.com>
+* Lorenzo Battistini <lorenzo.battistini@agilebg.com>
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * Carlos Dauden
+  * Jairo Llopis
+* Mourad EL HADJ MIMOUNE <mourad.elhadj.mimoune@akretion.com>
