+++ custom-addons/purchase_sale_inter_company/readme/DESCRIPTION.rst
@@ -1,0 +1,8 @@
+This module is useful if there are multiple companies in the same Odoo database and those companies sell goods or services among themselves.
+It allows to create a sale order in company A from a purchase order in company B.
+
+Imagine you have company A and company B in the same Odoo database:
+
+* Company A purchase goods from company B.
+* Company A will create a purchase order with company B as supplier.
+* This module automate the creation of the sale order in company B with company A as customer.
