+++ custom-addons/purchase_sale_inter_company/static/description/icon.png
