+++ custom-addons/purchase_sale_inter_company/tests/inter_company_purchase_sale.xml
@@ -1,0 +1,137 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <!--
+      © 2016-2018 Akretion France (www.akretion.com)
+      © 2016-2018 Tecnativa
+      License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+    -->
+    <!--
+        Create warehouse in Company A and Company B
+    -->
+    <record id="location_stock_company_a" model="stock.location">
+        <field name="name">Stock - A</field>
+        <field name="usage">internal</field>
+        <field name="company_id" ref="company_a" />
+    </record>
+    <record id="location_output_company_a" model="stock.location">
+        <field name="name">Output - A</field>
+        <field name="usage">internal</field>
+        <field name="company_id" ref="company_a" />
+    </record>
+    <record id="warehouse_company_a" model="stock.warehouse">
+        <field name="name">purchase warehouse - A</field>
+        <field name="code">CMPA</field>
+        <field name="wh_input_stock_loc_id" ref="location_stock_company_a" />
+        <field name="lot_stock_id" ref="location_stock_company_a" />
+        <field name="wh_output_stock_loc_id" ref="location_output_company_a" />
+        <field name="partner_id" ref="partner_company_a" />
+        <field name="company_id" ref="company_a" />
+    </record>
+    <record id="location_stock_company_b" model="stock.location">
+        <field name="name">Stock - B</field>
+        <field name="usage">internal</field>
+        <field name="company_id" ref="company_b" />
+    </record>
+    <record id="location_output_company_b" model="stock.location">
+        <field name="name">Output - B</field>
+        <field name="usage">internal</field>
+        <field name="company_id" ref="company_b" />
+    </record>
+    <record id="warehouse_company_b" model="stock.warehouse">
+        <field name="name">purchase warehouse - B</field>
+        <field name="code">CMPB</field>
+        <field name="wh_input_stock_loc_id" ref="location_stock_company_b" />
+        <field name="lot_stock_id" ref="location_stock_company_b" />
+        <field name="wh_output_stock_loc_id" ref="location_output_company_b" />
+        <field name="partner_id" ref="partner_company_b" />
+        <field name="company_id" ref="company_b" />
+    </record>
+    <!--
+        Apply warehouse in Company A and Company B
+        -->
+    <record id="company_a" model="res.company">
+        <field name="warehouse_id" ref="warehouse_company_a" />
+        <field name="sale_auto_validation">1</field>
+    </record>
+    <record id="company_b" model="res.company">
+        <field name="warehouse_id" ref="warehouse_company_b" />
+        <field name="sale_auto_validation">1</field>
+    </record>
+    <!--
+        Update User A and User B with purchase group
+        -->
+    <record id="user_company_a" model="res.users">
+        <field
+            name="groups_id"
+            eval="[(6, 0, [ref('account.group_account_manager'), ref('base.group_partner_manager'), ref('sales_team.group_sale_manager'), ref('purchase.group_purchase_user')])]"
+        />
+    </record>
+    <record id="user_company_b" model="res.users">
+        <field
+            name="groups_id"
+            eval="[(6, 0, [ref('account.group_account_manager'), ref('base.group_partner_manager'), ref('sales_team.group_sale_manager'), ref('purchase.group_purchase_user')])]"
+        />
+    </record>
+    <!--
+        Create inter company product
+        -->
+    <record id="product_consultant_multi_company" model="product.product">
+        <field name="name">Service Multi Company</field>
+        <field name="uom_id" ref="uom.product_uom_hour" />
+        <field name="uom_po_id" ref="uom.product_uom_hour" />
+        <field name="categ_id" ref="product.product_category_3" />
+        <field name="type">service</field>
+        <field name="company_id" ref="" />
+    </record>
+    <!--
+        Create draft purchase in Company A
+        -->
+    <record id="purchase_company_a" model="purchase.order">
+        <field name="partner_id" ref="partner_company_b" />
+        <field name="state">draft</field>
+        <field name="company_id" ref="company_a" />
+    </record>
+    <record id="purchase_line_company_a" model="purchase.order.line">
+        <field name="order_id" ref="purchase_company_a" />
+        <field name="product_id" ref="product_consultant_multi_company" />
+        <field name="product_uom" ref="uom.product_uom_hour" />
+        <field name="name">Service Multi Company</field>
+        <field
+            name="date_planned"
+            eval="(datetime.now()).strftime('%Y-%m-%d %H:%M:%S')"
+        />
+        <field name="price_unit" eval="450.0" />
+        <field name="product_qty" eval="3.0" />
+        <field name="company_id" ref="company_a" />
+    </record>
+    <record id="sequence_sale_journal_company_b" model="ir.sequence">
+        <field name="name">Account Sales Journal Company B</field>
+        <field eval="3" name="padding" />
+        <field name="prefix">SAJ-B/%(year)s/</field>
+        <field name="company_id" ref="company_b" />
+    </record>
+    <record id="sequence_purchase_journal_company_a" model="ir.sequence">
+        <field name="name">Account Expenses Journal Company A</field>
+        <field eval="3" name="padding" />
+        <field name="prefix">EXJ-A/%(year)s/</field>
+        <field name="company_id" ref="company_a" />
+    </record>
+    <record id="sales_journal_company_b" model="account.journal">
+        <field name="name">Sales Journal - (Company B)</field>
+        <field name="code">SAJ-B</field>
+        <field name="type">sale</field>
+        <field name="sequence_id" ref="sequence_sale_journal_company_b" />
+        <field name="default_credit_account_id" ref="a_sale_company_b" />
+        <field name="default_debit_account_id" ref="a_sale_company_b" />
+        <field name="company_id" ref="company_b" />
+    </record>
+    <record id="purchases_journal_company_a" model="account.journal">
+        <field name="name">Purchases Journal - (Company A)</field>
+        <field name="code">EXJ-A</field>
+        <field name="type">purchase</field>
+        <field name="sequence_id" ref="sequence_purchase_journal_company_a" />
+        <field name="default_credit_account_id" ref="a_expense_company_a" />
+        <field name="default_debit_account_id" ref="a_expense_company_a" />
+        <field name="company_id" ref="company_a" />
+    </record>
+</odoo>
