+++ custom-addons/purchase_sale_inter_company/views/res_config_view.xml
